+++ filter/class.tx_rnbase_filter_BaseFilter.php
@@ -26,6 +26,15 @@
 tx_rnbase::load('tx_rnbase_util_SearchBase');
 tx_rnbase::load('Tx_Rnbase_Database_Connection');
 
+/**
+ * tx_rnbase_IFilter
+ *
+ * @package         TYPO3
+ * @subpackage      rn_base
+ * @author          Rene Nitzsche <rene@system25.de>
+ * @license         http://www.gnu.org/licenses/lgpl.html
+ *                  GNU Lesser General Public License, version 3 or later
+ */
 interface tx_rnbase_IFilter
 {
     /**
@@ -57,24 +66,72 @@ interface tx_rnbase_IFilter
      */
     public function isSpecialSearch();
 }
+
+/**
+ * tx_rnbase_IFilterMarker
+ *
+ * @package         TYPO3
+ * @subpackage      rn_base
+ * @author          Rene Nitzsche <rene@system25.de>
+ * @license         http://www.gnu.org/licenses/lgpl.html
+ *                  GNU Lesser General Public License, version 3 or later
+ */
 interface tx_rnbase_IFilterMarker
 {
     public function parseTemplate($template, &$formatter, $confId, $marker = 'FILTER');
 }
 
+/**
+ * tx_rnbase_filter_BaseFilter
+ *
+ * @package         TYPO3
+ * @subpackage      rn_base
+ * @author          Rene Nitzsche <rene@system25.de>
+ * @author          Hannes Bochmann <hannes.bochmann@dmk-ebusiness.de>
+ * @license         http://www.gnu.org/licenses/lgpl.html
+ *                  GNU Lesser General Public License, version 3 or later
+ */
 class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilterMarker
 {
+
+    /**
+     * @var Tx_Rnbase_Configuration_ProcessorInterface
+     */
     private $configurations;
+
+    /**
+     * @var tx_rnbase_IParameters
+     */
     private $parameters;
+
+    /**
+     * @var string
+     */
     private $confId;
+
+    /**
+     * @var array
+     */
     protected $filterItems;
 
+    /**
+     * @var null|boolean wenn $doSearch auf null steht wird der return Wert von initFilter()
+     * in init() zurück gegeben. Ansonsten der Wert von $doSearch, dieser hat also Vorrang.
+     */
+    protected $doSearch = null;
+
+    /**
+     * @param tx_rnbase_IParameters $parameters
+     * @param Tx_Rnbase_Configuration_ProcessorInterface $configurations
+     * @param string $confId
+     */
     public function __construct(&$parameters, &$configurations, $confId)
     {
         $this->configurations = $configurations;
         $this->parameters = $parameters;
         $this->confId = $confId;
     }
+
     /**
      * Liefert das Config-Objekt
      *
@@ -84,6 +141,7 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
     {
         return $this->configurations;
     }
+
     /**
      * Liefert die Parameter
      *
@@ -93,6 +151,7 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
     {
         return $this->parameters;
     }
+
     /**
      * Liefert die Basis-ConfigId. Diese sollte immer mit einem Punkt enden: myview.
      *
@@ -102,6 +161,7 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
     {
         return $this->confId;
     }
+
     /**
      * Abgeleitete Filter können diese Methode überschreiben und zusätzlich Filter setzen
      *
@@ -112,12 +172,13 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
     public function init(&$fields, &$options)
     {
         tx_rnbase_util_SearchBase::setConfigFields($fields, $this->getConfigurations(), $this->getConfId().'fields.');
-        // Optionen
         tx_rnbase_util_SearchBase::setConfigOptions($options, $this->getConfigurations(), $this->getConfId().'options.');
 
         $fields = $this->handleSysCategoryFilter($fields);
 
-        return $this->initFilter($fields, $options, $this->getParameters(), $this->getConfigurations(), $this->getConfId());
+        return $this->shouldSearchBeDone(
+            $this->initFilter($fields, $options, $this->getParameters(), $this->getConfigurations(), $this->getConfId())
+        );
     }
 
     /**
@@ -126,25 +187,29 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
      */
     protected function handleSysCategoryFilter(array $fields)
     {
-        if ($this->getConfigurations()->get($this->getConfId() . 'useSysCategoriesOfItemFromParameters')) {
-            $fields = $this->getCategoryFilterUtility()->setFieldsBySysCategoriesOfItemFromParameters(
-                $fields, $this->getConfigurations(),
-                $this->getConfId() . 'useSysCategoriesOfItemFromParameters.'
-            );
-        }
+        $typoScriptPathsToFilterUtilityMethod = array(
+            'useSysCategoriesOfItemFromParameters' => 'setFieldsBySysCategoriesOfItemFromParameters',
+            'useSysCategoriesOfContentElement' => 'setFieldsBySysCategoriesOfContentElement',
+            'useSysCategoriesFromParameters' => 'setFieldsBySysCategoriesFromParameters',
+        );
 
-        if ($this->getConfigurations()->get($this->getConfId() . 'useSysCategoriesOfContentElement')) {
-            $fields = $this->getCategoryFilterUtility()->setFieldsBySysCategoriesOfContentElement(
-                $fields, $this->getConfigurations(),
-                $this->getConfId() . 'useSysCategoriesOfContentElement.'
-            );
-        }
+        foreach ($typoScriptPathsToFilterUtilityMethod as $typoScriptPath => $filterUtilityMethod) {
+            if ($this->getConfigurations()->get($this->getConfId() . $typoScriptPath)) {
+                $fieldsBefore = $fields;
+                $fields = $this->getCategoryFilterUtility()->$filterUtilityMethod(
+                    $fields, $this->getConfigurations(),
+                    $this->getConfId() . $typoScriptPath . '.'
+                );
 
-        if ($this->getConfigurations()->get($this->getConfId() . 'useSysCategoriesFromParameters')) {
-            $fields = $this->getCategoryFilterUtility()->setFieldsBySysCategoriesFromParameters(
-                $fields, $this->getConfigurations(),
-                $this->getConfId() . 'useSysCategoriesFromParameters.'
-            );
+                if (
+                    $this->getConfigurations()->get($this->getConfId() . $typoScriptPath . '.dontSearchIfNoCategoriesFound') &&
+                    // wenn sich die $fields nicht geändert haben, dann wurden keine Kategorie
+                    // gefunden.
+                    $fieldsBefore == $fields
+                ) {
+                    $this->doSearch = false;
+                }
+            }
         }
 
         return $fields;
@@ -156,6 +221,20 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
     protected function getCategoryFilterUtility()
     {
         return tx_rnbase::makeInstance('Tx_Rnbase_Category_FilterUtility');
+    }
+
+    /**
+     * @param boolean $fallback
+     * @return boolean
+     */
+    protected function shouldSearchBeDone($fallback)
+    {
+        $doSearch = $this->doSearch;
+        if ($doSearch === null) {
+            $doSearch = $fallback;
+        }
+
+        return $doSearch;
     }
 
     /**
@@ -181,6 +260,7 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
     {
         return true;
     }
+
     /**
      * Hilfsmethode zum Setzen von Filtern aus den Parametern. Ein schon gesetzter Wert im Field-Array
      * wird nicht überschrieben. Die
@@ -201,10 +281,14 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
         }
     }
 
+    /**
+     * @param tx_rnbase_IFilterItem $item
+     */
     public function addFilterItem(tx_rnbase_IFilterItem $item)
     {
         $this->filterItems[] = $item;
     }
+
     /**
      * Returns all filter items set.
      *
@@ -214,6 +298,7 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
     {
         return $this->filterItems;
     }
+
     /**
      * Fabrikmethode zur Erstellung von Filtern. Die Klasse des Filters kann entweder direkt angegeben werden oder
      * wird über die Config gelesen. Klappt beides nicht, wird der Standardfilter geliefert.
@@ -248,10 +333,15 @@ class tx_rnbase_filter_BaseFilter implements tx_rnbase_IFilter, tx_rnbase_IFilte
         return is_array($this->inputData) && count($this->inputData);
     }
 
+    /**
+     * {@inheritDoc}
+     * @see tx_rnbase_IFilter::getMarker()
+     */
     public function getMarker()
     {
         return $this;
     }
+
     /**
      * Liefert einfach das Template zurück. Ein echter FilterMarker hat hier die Möglichkeit sein
      * Such-Formular in das HTML-Template zu schreiben.